pyproject.toml
[build-system]
requires = ["setuptools>=63.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ccsd-gpu"
version = "0.1.0"
description = "Standalone GPU-accelerated CCSD gradient prototype"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy",
    "pyscf>=2.8.0",
    "cutensor-cu12==2.2.0",
    "cupy-cuda12x==13.4.1",
    "gpu4pyscf-cuda12x>=1.3.1",
    "nvidia-cublas-cu12>=12.9.2.10",
    "nvidia-cusolver-cu12>=11.7.5.82",
    "nvidia-cuda-nvrtc-cu12>=12.9.86",
    "nvidia-nvjitlink-cu12>=12.9.86",
    "packaging>=26.0",
    "nvidia-cuda-runtime-cu12>=12.9.79",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
ccsd_gpu = []
"ccsd_gpu.cuda" = ["rys_data_n9.json"]
